test: optimized fdp initialization
