+++ test/command/pod.spec.ts
@@ -1,29 +1,17 @@
 import { describeCommand, invokeTestCli } from '../utility'
-import { createFdp, getRandomString } from '../utils'
+import { getRandomString } from '../utils'
+import { createFdpAndImport } from '../utility/fdp'
 
 describeCommand(
   'Test Pod command',
   ({ consoleMessages }) => {
     it('should list pods', async () => {
-      const account = getRandomString()
-      const accountPassword = getRandomString()
       const podName1 = getRandomString()
       const podName2 = getRandomString()
-      const fdp = createFdp()
+      const { fdp, account, accountPassword } = await createFdpAndImport()
 
-      const wallet = fdp.account.createWallet()
       await fdp.personalStorage.create(podName1)
       await fdp.personalStorage.create(podName2)
-
-      await invokeTestCli([
-        'account',
-        'import',
-        wallet.mnemonic.phrase,
-        '--name',
-        account,
-        '--password',
-        accountPassword,
-      ])
       consoleMessages.length = 0
 
       await invokeTestCli(['pod', 'list', '--account', account, '--password', accountPassword])
@@ -34,24 +22,11 @@ describeCommand(
     })
 
     it('should create pods', async () => {
-      const account = getRandomString()
-      const accountPassword = getRandomString()
       const podName1 = getRandomString()
       const podName2 = getRandomString()
-      const fdp = createFdp()
+      const { fdp, account, accountPassword } = await createFdpAndImport()
 
-      const wallet = fdp.account.createWallet()
-      await invokeTestCli([
-        'account',
-        'import',
-        wallet.mnemonic.phrase,
-        '--name',
-        account,
-        '--password',
-        accountPassword,
-      ])
       consoleMessages.length = 0
-
       await invokeTestCli(['pod', 'create', podName1, '--account', account, '--password', accountPassword])
       expect(consoleMessages[0]).toContain('Pod created successfully!')
       consoleMessages.length = 0
@@ -73,23 +48,9 @@ describeCommand(
     })
 
     it('should delete pods', async () => {
-      const account = getRandomString()
-      const accountPassword = getRandomString()
       const podName1 = getRandomString()
       const podName2 = getRandomString()
-      const fdp = createFdp()
-
-      const wallet = fdp.account.createWallet()
-      await invokeTestCli([
-        'account',
-        'import',
-        wallet.mnemonic.phrase,
-        '--name',
-        account,
-        '--password',
-        accountPassword,
-      ])
-      consoleMessages.length = 0
+      const { fdp, account, accountPassword } = await createFdpAndImport()
 
       await invokeTestCli(['pod', 'create', podName1, '--account', account, '--password', accountPassword])
       await invokeTestCli(['pod', 'create', podName2, '--account', account, '--password', accountPassword])
